added rewards to dashboard + ui fixes
