+++ packages/ui/app/_components/markets/PoolRows.tsx
@@ -3,20 +3,17 @@
 import { type Reward } from '@ionicprotocol/types';
 import Link from 'next/link';
 import { useSearchParams } from 'next/navigation';
-import { useState, type Dispatch, type SetStateAction } from 'react';
-import { formatEther, type Address } from 'viem';
-import { base } from 'viem/chains';
-import { useChainId } from 'wagmi';
+import { type Dispatch, type SetStateAction } from 'react';
+import { type Address } from 'viem';
 
 import { getAssetName } from '../../util/utils';
 import ConnectButton from '../ConnectButton';
 import { PopupMode } from '../popup/page';
-import ResultHandler from '../ResultHandler';
+
+import Rewards from './Rewards';
 
 import { pools } from '@ui/constants/index';
 import { useMultiIonic } from '@ui/context/MultiIonicContext';
-import { useSdk } from '@ui/hooks/ionic/useSdk';
-import { useAssetClaimableRewards } from '@ui/hooks/rewards/useAssetClaimableRewards';
 import type { LoopMarketData } from '@ui/hooks/useLoopMarkets';
 import type { MarketData } from '@ui/types/TokensDataMap';
 import { multipliers } from '@ui/utils/multipliers';
@@ -168,7 +165,7 @@ const PoolRows = ({
       </Link>
 
       <h3
-        className={` col-span-2 flex md:block justify-between md:justify-center px-2 md:px-0 text-xs  md:py-4 py-0 items-center mb-2 md:mb-0 `}
+        className={` col-span-2 flex  justify-between md:justify-center px-2 md:px-0 text-xs  md:py-4 py-0 items-center mb-2 md:mb-0 `}
       >
         <span className="text-white/40 font-semibold mr-2 md:hidden text-left  ">
           SUPPLY APR:
@@ -521,90 +518,6 @@ const PoolRows = ({
   );
 };
 
-const REWARDS_TO_SYMBOL: Record<number, Record<Address, string>> = {
-  [base.id]: {
-    '0x3eE5e23eEE121094f1cFc0Ccc79d6C809Ebd22e5': 'ION'
-  }
-};
-
-type RewardsProps = {
-  cToken: Address;
-  pool: Address;
-  poolChainId: number;
-};
-const Rewards = ({ cToken, pool, poolChainId }: RewardsProps) => {
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const { data: rewardsData } = useAssetClaimableRewards(
-    cToken,
-    pool,
-    poolChainId
-  );
-  const chainId = useChainId();
-  const sdk = useSdk(poolChainId);
-
-  const claimRewards = async () => {
-    try {
-      const result = await handleSwitchOriginChain(poolChainId, chainId);
-      if (result) {
-        setIsLoading(true);
-        const tx = await sdk?.claimRewardsForMarket(
-          cToken,
-          rewardsData?.map((r) => r.flywheel!) ?? []
-        );
-        setIsLoading(false);
-        console.warn('claim tx: ', tx);
-      }
-    } catch (err) {
-      setIsLoading(false);
-      console.warn(err);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const totalRewards =
-    rewardsData?.reduce((acc, reward) => acc + reward.amount, 0n) ?? 0n;
-
-  return (
-    <div className="pb-4">
-      {rewardsData?.map((rewards, index) => (
-        <div
-          className="flex"
-          key={index}
-        >
-          <img
-            alt=""
-            className="size-4 rounded mr-1"
-            src={`/img/symbols/32/color/${REWARDS_TO_SYMBOL[poolChainId][rewards.rewardToken].toLowerCase()}.png`}
-          />{' '}
-          +{' '}
-          {Number(formatEther(rewards.amount)).toLocaleString('en-US', {
-            maximumFractionDigits: 1
-          })}{' '}
-          {REWARDS_TO_SYMBOL[poolChainId][rewards.rewardToken]}
-        </div>
-      ))}
-      {totalRewards > 0n && (
-        <div className="flex justify-center pt-1">
-          <button
-            className={`rounded-md bg-accent text-black py-1 px-3 uppercase truncate `}
-            onClick={claimRewards}
-          >
-            <ResultHandler
-              isLoading={isLoading}
-              height="20"
-              width="20"
-              color={'#000000'}
-            >
-              Claim Rewards
-            </ResultHandler>
-          </button>
-        </div>
-      )}
-    </div>
-  );
-};
-
 type BorrowPopoverProps = {
   dropdownSelectedChain: number;
   borrowAPR?: number;
@@ -663,7 +576,7 @@ const BorrowPopover = ({
             ?.flywheel && (
             <Rewards
               cToken={cToken}
-              pool={pool}
+              comptrollerAddress={pool}
               poolChainId={dropdownSelectedChain}
             />
           )}
